backend/requirements.txt
django
django-esi
django-eve-auth
